refactor(amethyst): remove legacy project naming
