+++ apps/amethyst/app/(tabs)/profile/[handle].tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react";
 import { ActivityIndicator, Image, View } from "react-native";
 import { Link, Stack, useLocalSearchParams } from "expo-router";
-import PlayFeedCard from "@/components/songish/PlayFeedCard";
-import RightRail from "@/components/songish/RightRail";
-import SongishShell, {
+import PlayFeedCard from "@/components/teal/PlayFeedCard";
+import RightRail from "@/components/teal/RightRail";
+import TealShell, {
   SectionHeading,
-} from "@/components/songish/SongishShell";
+} from "@/components/teal/TealShell";
 import { Button } from "@/components/ui/button";
 import { Text } from "@/components/ui/text";
 import { resolveHandle } from "@/lib/atp/pid";
@@ -91,7 +91,7 @@ export default function ProfileScreen() {
   const isSelf = did === pdsAgent?.did;
 
   return (
-    <SongishShell rightRail={<RightRail />}>
+    <TealShell rightRail={<RightRail />}>
       <Stack.Screen
         options={{ title: actor || "Profile", headerShown: false }}
       />
@@ -180,6 +180,6 @@ export default function ProfileScreen() {
           )}
         </>
       )}
-    </SongishShell>
+    </TealShell>
   );
 }
